+++ absl/testing/tests/absltest_filtering_test_helper.py
@@ -23,6 +23,7 @@ import sys
 
 from absl import app
 from absl.testing import absltest
+from absl.testing import parameterized
 
 
 class ClassA(absltest.TestCase):
@@ -56,6 +57,21 @@ class ClassB(absltest.TestCase):
   def testE(self):
     sys.stderr.write('\nclass B test E\n')
     self.fail('Force failure')
+
+
+class ParameterizedTest(parameterized.TestCase):
+  """Helper parameterized test case for absltest_filtering_test."""
+
+  @parameterized.parameters([1, 2])
+  def test_unnamed(self, value):
+    sys.stderr.write('\nparameterized unnamed %s' % value)
+
+  @parameterized.named_parameters(
+      ('test1', 1),
+      ('test2', 2),
+  )
+  def test_named(self, value):
+    sys.stderr.write('\nparameterized named %s' % value)
 
 
 def main(argv):